+++ module/Rest/src/Action/Visit/TagVisitsAction.php
@@ -7,7 +7,7 @@ namespace Shlinkio\Shlink\Rest\Action\Visit;
 use Laminas\Diactoros\Response\JsonResponse;
 use Psr\Http\Message\ResponseInterface as Response;
 use Psr\Http\Message\ServerRequestInterface as Request;
-use Shlinkio\Shlink\Common\Paginator\Util\PaginatorUtilsTrait;
+use Shlinkio\Shlink\Common\Paginator\Util\PagerfantaUtilsTrait;
 use Shlinkio\Shlink\Core\Model\VisitsParams;
 use Shlinkio\Shlink\Core\Service\VisitsTrackerInterface;
 use Shlinkio\Shlink\Rest\Action\AbstractRestAction;
@@ -15,7 +15,7 @@ use Shlinkio\Shlink\Rest\Middleware\AuthenticationMiddleware;
 
 class TagVisitsAction extends AbstractRestAction
 {
-    use PaginatorUtilsTrait;
+    use PagerfantaUtilsTrait;
 
     protected const ROUTE_PATH = '/tags/{tag}/visits';
     protected const ROUTE_ALLOWED_METHODS = [self::METHOD_GET];